cfs cmd
